refactor: add ProductRequest and make Product use Price

+++ src/test/java/kitchenpos/domain/PriceTest.java
@@ -1,6 +1,5 @@
 package kitchenpos.domain;
 
-import java.math.BigDecimal;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 
@@ -15,7 +14,7 @@ class PriceTest {
     @Test
     void 가격이_0보다_작으면_에러가_발생해야_한다() {
         // given
-        final BigDecimal invalidPrice = BigDecimal.valueOf(-1);
+        final Long invalidPrice = -1L;
 
         // when and then
         Assertions.assertThatThrownBy(() -> new Price(invalidPrice))
